Second attempt to automatically close positions on Kraken
It appears that Kraken actually modifies the volume of open positions over time, and requires that the volume match exactly when you place an order to close a position. Arbitrader remembers the volume from when it opened the position, but if the position was open for a few days it may have changed. We're going to remember the order ID and fetch the new volume (if the exchange supports it; not all of them do) and use that to close the position. If anything goes wrong we'll fall back to the volume we used to open the position.

In small scale, short term tests this strategy appears to work. I'm still waiting on a longer term position to close so I can see it "in the wild".