Corralling complexity, one iteration at a time
XChange is a great library but it's very complex and has a lot of final classes that can't be mocked. Exchanges are huge object graphs with an enormous number of permutations, and to test this beast we're going to need to generate a lot of different Exchanges in different configurations. A Builder is going to go a long way toward making that a little more pleasant.