Configurable slow ticker warning
My logs are filling up with slow ticker warnings and I'd like to tweak it, but because of all the various factors the "correct" value isn't going to be the same for everyone. So I made it something you can change in your `application.yaml` to make your logs just the right amount of noisy.