Remind the user of the total USD balance from the start of a trade at the end of the trade
It's not a complicated feature but it will make it so you don't have to scroll back up through hundreds of lines of logs to double check how well a trade really did. It will show you what the balance was when the trades were opened, then on the next line will be the current balance so you can easily see the difference. Having this little bit of extra information will help to debug the estimated profit calculation, which seems to be wrong a lot.