Wrong scales
The scales on the CurrencyPairs are for incoming prices. We want the scale for the coin (eg. 8 decimal places for BTC) which is on the Currency metadata instead. Some exchanges seem to be missing metadata for some pairs, so we need to wrap it up in an Optional to let it fall back to our constants if we can't look it up.