trading blackout windows
Allows a user to create a file named 'blackout' in the current working directory with "extended" ISO-8601 dates indicating that the bot should not trade during the specified times. Each line of the file should look like this:

Cool Coins,2011-12-03T10:15:30+01:00,2011-12-03T12:15:30+01:00

The above indicates that for the exchange "Cool Coins" from 10:15 to 12:15 on Dec 3, 2011 we should not trade. The file can have as many lines as you wish and may safely be changed or deleted while the bot is running. Each time the bot wants to trade, it will attempt to read the file and check whether the current time is within any of the specified blackout windows.