optional timeout for how long an open position may remain open
This is a good option to limit the effect of ongoing fees from the exchange where you have your short position. If you leave that position open long enough it can easily eat up all the profit from your trades, and currently Arbitrader is not capable of tracking those costs.