force a scale with more decimals for spread computation
Some exchanges return prices with very few decimal places which causes the spread calculation to toggle between 0.00 and 0.01. When the entry threshold is a number with more decimals like 0.008 the lack of granularity makes the bot thrash between wanting to enter and wanting to exit. This change will force the prices to have more decimal places during spread computation and give us more granular spread values to work with.