allow null tickers during initialization
We may not receive messages over websocket immediately, so we need to not disqualify exchanges and trading pairs that don't have ticker values yet. They will be skipped as needed during the tick until they have received data. The alternative would be to have a blocking wait for each streaming exchange and currency pair until we got the first message, but it would would take an undetermined amount of time.