switch to Amazon Corretto JDK 8 base images
There is no technical reason that makes this change necessary. It's more of a forward looking change because I trust that Amazon is eating its own dog food with Corretto (for instance, it's the AWS Lambda runtime) and has a profit motive to keep maintaining it into the future. I'm not nearly as confident that "adoptopenjdk" is going to be around in the long term.

Sticking with JDK 8 for now, but 11 is on the horizon.