pass custom arguments to `getTicker()`
The MarketDataService#getTicker() method has a vararg parameter that is mostly unused, and Arbitrader has simply omitted it up to this point for all exchanges. Unfortunately the streaming MarketDataService for Gemini has a bug where a default value is zero and literally any value greater than zero would solve our problem. If we could just pass a "1" to its `getTicker()` method...

This feature allows a YAML list in your configuration file of constant values to be passed into this `getTicker()` method for a given exchange. It allows a clean workaround of this Gemini bug and I think it may be useful for working around other similar bugs in the future. While the "correct" solution is probably to change the default value in XChange (and I plan to submit a PR for that too) this feature will allow working around it *right now* instead of waiting for the next release of XChange.