unify active position
Looks like the ActivePosition got split across TradingService and TradingScheduler which is making the behavior of tracking active positions act strangely, especially if you have quit and restarted the program, because it is being tracked independently in two places. This PR moves ownership of the ActivePosition over to TradingService and makes TradingScheduler access it via getter and setter.