first run at effective prices
This isn't well tested so apologies if it doesn't work, but I think the math is at least pretty close to right. It's a "simple" version where we incorporate the fees in both the entry and exit spreads, although Allirion has some fancier ideas that might turn out to work better. From what I can tell so far it pushes the spreadIn down and the spreadOut up which is what I would expect. It seems much harder to get into trades and it will also be harder to get back out again. I'll be running this for a few days to see how it works and keep iterating on it but I wanted to share and get some more eyes on it.