ignore crypto balances below the exchange's minimum trade size
I had a balance of 0.00000001 BTC in BitFlyer and this prevented me from starting the bot. I can't trade less than 0.001 BTC on BitFlyer so there is no way to exactly zero out the balance without first adding more funds. Instead, I'd like the bot to simply ignore that tiny amount. The intention of this check was to stop the bot from starting up when there were "large" positions; trades that didn't close properly or a user's long term investment stash and they accidentally started the bot on that account.